Add margin-bottom and min-width to improve layout
A margin-bottom of 2 units was added to the div within the PlotCreationController.tsx to provide better spacing between elements. Likewise, a minimum width was added to the div within PlotInfoForm.tsx to ensure that the PricePerM2 component has sufficient space for displaying its content without breaking the layout. These changes improvise the UI visual balance and prevent potential layout issues.