Update README with new error handling for file renaming
Added documentation in README regarding a new error message that can appear in Browser DevTools when a file is renamed. This edit is meant to guide developers in understanding the error and the solution, which is to simply ignore it. This information has been added in order to improve the troubleshooting process and increase efficiency among developers.