Extend utilities in Tailwind CSS configuration
New utility classes '.dragNDrop', '.plotImg', and '.plotForm' were added to the tailwind.config.js file to handle specific interface elements in the application. The '.highlight::after' CSS class was relocated from the global.css file to the Tailwind configuration to unify the manageability of styles. This particularly fosters a cleaner approach to styling by reducing dependencies on global CSS.