Refine PointFromTextInput type for plot creation
This commit modifies the definition of the PointFromTextInput type in store/draftPlot/common.ts file. Instead of manually defining the lat and lng properties as number or string, a more generic approach is used. Utilizing type-fest's Simplify feature, we are now associating LatKey and LngKey with the same data types. This change contributes to cleaner and more maintainable code by eliminating redundancy and enhancing type safety.